% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_namelists.R
\name{namelists}
\alias{namelists}
\title{Available nameslists}
\description{
\itemize{
    \item docker: Lists of names used for docker. Stolen from \url{https://github.com/moby/moby/blob/master/pkg/namesgenerator/names-generator.go}
}
}
