% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_names.R
\name{random_names}
\alias{random_names}
\title{Get random names}
\usage{
random_names(n = 1, namelist = "docker", wordsep = "_", alliterate = FALSE)
}
\arguments{
\item{n}{Number of names needed. If n is greater than number of possible
word combinations, a number will be added as a suffix.}

\item{namelist}{A name list. Either a character that corresponds to one of
the name lists included in the package or a list with character arrays that
includes custom words to be combined. Custom word lists can have as many
character vectors as desired. A word from each character list will be chose
randomly. Name of the elements are not used. See \code{\link{?namelists}} for
a list of available name lists.}

\item{wordsep}{Character to be used when separating the words}

\item{alliterate}{}
}
\value{
A character vector of length n.
}
\description{
Get random names
}
\examples{
random_names(n=4, namelist = 'docker')
}
