% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_namelists.R
\name{namelists}
\alias{namelists}
\title{Available nameslists}
\description{
\itemize{
    \item docker: Lists of names used for docker containers. Follows the adjective_surname formula.
    Stolen from \url{https://github.com/moby/moby/blob/master/pkg/namesgenerator/names-generator.go}
    23400 possible combinations
    \item xcom2_names: Xcom2 mission names following adjective_noun formula.
    14351 possible combinations
    \item xcom2_words: Xcom2 mission names following noun_noun formula.
    1617 possible combinations
}
}
